+++ src/components/FrontCardView.tsx
